Double precision for FOV, to hopefully avoid the large incidence of OOS I now see when running server recordings locally

+++ brogue-1.7.5/src/brogue/IO.c
@@ -1640,7 +1640,7 @@ short adjustedLightValue(short x) {
     if (x <= LIGHT_SMOOTHING_THRESHOLD) {
         return x;
     } else {
-        return (short) (sqrt(((float) x)/LIGHT_SMOOTHING_THRESHOLD)*LIGHT_SMOOTHING_THRESHOLD + FLOAT_FUDGE);
+        return (short) (sqrt(((double) x)/LIGHT_SMOOTHING_THRESHOLD)*LIGHT_SMOOTHING_THRESHOLD + FLOAT_FUDGE);
     }
 }
